fix issue #7676: wrong handler around MULTIEXPR
Citus was checking for presence of sublink, but forgot to manage
multiexpr while evaluating clauses during planning.
At this stage (citus planner), it's not always possible to call PostgreSQL
code because the tree is not yet ready for PostgreSQL pure executor.

https://github.com/citusdata/citus/issues/7676

Fixed by adding a new function to check sublink or multiexp in the tree.