Manage special case for "GRANT ALL" with columns
`GRANT/REVOKE ALL (col, ...)` cannot be mixed with other privilege and
because there is at least one column defined, the privileges node is
defined so the existing code is valid for table but not for columns.

Add a condition to ensure `ALL` is defined alone and added to the
privilege string when needed.

Comment added based on:
https://github.com/citusdata/citus/pull/7918#discussion_r1994929886