Manage special case for "ALL" with columns
`ALL (col, ...)` cannot be mixed with other privilege and because there
is at least one column defined, the privileges node is defined so the
existing code is valid for table but not for columns.

Add a condition to ensure `ALL` is defined alone and added to the
privilege string when needed.