Fix ObjectClass declaration for PG17 since it was removed
Relevant PG commit:
https://github.com/postgres/postgres/commit/89e5ef7e21812916c9cf9fcf56e45f0f74034656
89e5ef7e21812916c9cf9fcf56e45f0f74034656

We had already provided a fix for this in the following commit
https://github.com/citusdata/citus/commit/da2624cee82acd1ab53435b215fbb67c190a9e96

However, this solution wasn't enough for the commits on main.
Specifically, we had issues with the following commit:
https://github.com/citusdata/citus/commit/1d55debb988d3e630205467f6e2e903294a5eb5d
Problem: https://github.com/citusdata/citus/actions/runs/13806825532/attempts/1#summary-38619483894

This new solution is better anyway. We define exactly what was
previously defined in PG<17.