Change the way ACL are checked whe locking shard
In `EnsureTablePermissions()`, which is used in `lock_shard_resources()`
but also in few other places, citus does check the ACL at the table
level but is not examining attributes ACL.
It prevents GRANT INSERT/UPDATE(col) to work as citus requires to lock
shared but this GRANT is not a table lock.

Change the behavior to check attributes ACL also.

Other usage:

* `Datum get_shard_id_for_distribution_column()`, looks safe.
* `Datum master_create_empty_shard()`, in this case it is apparently
  really required to be more careful. This is managed in the next
  commit.