Add finer control in EnsureTablePermission
Previously this function was only checking ACL at table level, however
for propagating GRANT on attributes, it is required to lock the shard
while the user may not have table privilege at all (but attribute level
ones).

In order to solve that, I add a new parameter to the function: the ACL
mask to be used when checking attributes acl:

-----
Check that the current user has `mode` permissions on relationId.

If not, also check relationId's attributes with `mask`, error out
privileges are not defined.

ACL mask is used because we assume that user has enought privilege
to distribute a table when either ACL_INSERT on the TABLE or
ACL_INSERT on ALL attributes.

In other situations, having a single attribute privilege is enough.
-----