Remove nodeid and groupid output in a test
In multi_read_from_secondaries, there was a select on nodeid and groupid.
But in another (future) test a node is dropped then added again, it's ok but
its nodeid and groupid are increased which leads to a faillure in this
test file.

Apparently nodeid and groupid are not really the subject of the test so
just remove them from the SELECT.