Factorize code from pg_get_table_grants
The code was suppopsed to be heavely based on aclexplode() from
PostgreSQL. It's partially true and I moved the essence of the part of
the code into a new static function to build GRANT sql strings.

Also removed the INDENT-OFF marker has the code has changed and should
anyway be reviewed to manage "grantor".