Add check to identify views converted to RTE_SUBQUERY in 15.13, 14.18
Previously, when views were converted to RTE_SUBQUERY the relid
would be cleared in PG15. In this patch of PG15, relid is retained.
Therefore, we add a check with the "relkind and rtekind" to
identify the converted views in 15.13

Part of https://github.com/citusdata/citus/commit/c98341e4ed3967219dd5ae0e264dc41a7f442f1c