PG18 - fix naming diffs of child FK constraints (#8226)
PG18 changed the names generated for child foreign key constraints.
https://github.com/postgres/postgres/commit/3db61db48

The test failures in Citus regression suite are all changing the name of
a constraint from `'sensors%'` to `'%to_parent%_1'`: the naming is very
nice here because `to_parent` means that we have a foreign key to a
parent table.

To fix the diff, we exclude those constraints from the output.
To verify correctness, we still count the problematic constraints to
make sure they are there - we are simply removing them from the first
output (we add this count query right after the previous one)

Fixes #8126

---------

Co-authored-by: naisila <nicypp@gmail.com>